Version:1.0 - Getting Individual Modules Working/requirements.txt
python-dotenv
openai
json
requests